fixed bug and clean code, format  code
